requirements.txt
python-multipart==0.0.6
webcolors==1.13
requests==2.31.0
uvicorn==0.24.0.post1
fastapi==0.104.1
Pillow==10.1.0